[Checkers Core] Checkers Logic
